finally segtree worksgit add *  what was wrong was testing (and surprise surprise, a boundary). The idea for lookups is the following, starting from root of segtree, everytime check is congruent (i.e. (i,j) such that i < j ) and whether you already found it, if not yet found, break the lookup into two calls: one for the left subtree and other for the right, but there's a catch.
whenever going to a left/right subtree, update segment looked as to fit the new boundaries you're entering. This cause the segment to be looked for may be covered by several (I think at least two) segments of the segtree. This can be achieved by fitting the left/right bound of the interval when going to the right/left subtree

Im happy :D